% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C150toC130.R
\name{C150toC130}
\alias{C150toC130}
\title{C150toC130}
\usage{
C150toC130(C150, speciesCode)
}
\arguments{
\item{C150}{the girth in cm measured at 150 cm (Belgian style)}

\item{speciesCode}{an integer defining species (see table eqNumber)}
}
\value{
the estimated girth in cm at 130 cm (international style)
}
\description{
Girth can be measured at different height of measurement (hom). This function
converts girth measured with hom = 150 cm into girth with hom = 130 cm
}
\examples{
C150toC130(C150 = 90, speciesCode = 1)
}
