% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEquationWithParam.R
\name{getEquationWithParam}
\alias{getEquationWithParam}
\title{getEquationWithParam}
\usage{
getEquationWithParam(tarif, speciesCode)
}
\arguments{
\item{tarif}{an integer, either 1, 2 or 3, refering to a tarif function(see the readme)}

\item{speciesCode}{an integer indicating species code}
}
\value{
a character showing the equation and the parameters
}
\description{
getEquationWithParam
}
