% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeciesCodes.R
\name{getSpeciesCodes}
\alias{getSpeciesCodes}
\title{getSpeciesCodes}
\usage{
getSpeciesCodes()
}
\value{
species table with names and codes
}
\description{
This function return a species table with names and codes.
}
\examples{
getSpeciesCodes()
}
