% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarif1.R
\name{tarif1}
\alias{tarif1}
\title{Predict tree volume from tree girth}
\usage{
tarif1(C130, speciesCode)
}
\arguments{
\item{C130}{A single value of girth in centimeter with hom = 130 cm}

\item{speciesCode}{An integer defining species (see table eqNumber)}
}
\value{
tree commercial volume (VC22) in cubic meters
}
\description{
Predict tree commercial volume (VC22) as a function of tree girth (C130) for a given species
}
\examples{
tarif1(90,1)

}
