% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarif2.R
\name{tarif2}
\alias{tarif2}
\title{Predict tree volume from tree girth and dominant height}
\usage{
tarif2(C130, hdom, speciesCode)
}
\arguments{
\item{C130}{A vector of girth in centimeter with hom = 130 cm}

\item{hdom}{A value of stand dominant height in meter}

\item{speciesCode}{An integer defining species (see table eqNumber)}
}
\value{
tree merchantable stem volume (VC22) in cubic meters
}
\description{
Predict tree commercial volume (VC22) as a function of tree girth (C130) and dominant height (m) for a given species
}
\examples{
tarif2(90,21,1)

}
