% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarif4.R
\name{tarif4}
\alias{tarif4}
\title{Predict merchandable branch volume from tree girth and height}
\usage{
tarif4(C130, speciesCode)
}
\arguments{
\item{C130}{A vector of girth in centimeter with hom = 130 cm}
}
\value{
tree merchantable branch volume (VC22) in cubic meters
}
\description{
Predict tree commercial volume of branches (VC22) as a function of tree girth (C130) and height (m) for a given species
}
\examples{
tarif4(90,1)

}
