% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarif3.R
\name{tarif3}
\alias{tarif3}
\title{Predict tree volume from tree girth and height}
\usage{
tarif3(C130, height, speciesCode)
}
\arguments{
\item{C130}{A vector of girth in centimeter with hom = 130 cm}

\item{height}{A vector of tree height in meter. It must be of the same length as C130.}

\item{speciesCode}{An integer defining species (see table eqNumber)}
}
\value{
tree merchantable stem volume (VC22) in cubic meters
}
\description{
Predict tree commercial volume (VC22) as a function of tree girth (C130) and height (m) for a given species
}
\examples{
tarif3(90,20,1)

}
